fix: close release verification gaps
