+++ .pre-commit-config.yaml
@@ -41,8 +41,30 @@ repos:
       # does not require optional contributor tooling.
       - id: cargo-clean
         name: cargo clean (clear build cache)
-        entry: scripts/sweep-build-artifacts.sh
+        entry: scripts/clear-build-cache.sh
         language: system
         pass_filenames: false
         stages: [pre-commit]
+        always_run: true
+
+      # Drop build artifacts the most recent build did not use.
+      #
+      # `target/` keeps superseded test and binary artifacts, re-linked on each
+      # change and never evicted, so it can grow without bound. This prunes what
+      # the build just finished with, so a fresh build's output survives and
+      # everything it superseded goes.
+      #
+      # `post-commit`, not `pre-commit`: `--stamp` must precede the build for
+      # `--file` to spare it, and the hooks above build. Running here means the
+      # stamp written last time bounds the artifacts written this time.
+      #
+      # Never fails a commit: `cargo-sweep` is optional tooling, and a missing
+      # binary or a locked target directory is not a reason to reject work.
+      - id: cargo-sweep
+        name: cargo sweep (drop superseded build artifacts)
+        entry: scripts/sweep-build-artifacts.sh
+        language: system
+        types: [rust]
+        pass_filenames: false
+        stages: [post-commit]
         always_run: true
